requirements.txt
requests==2.31.0
tweepy==4.14.0
praw==7.7.1
python-telegram-bot==20.6
pandas==2.1.3
matplotlib==3.8.2
python-dotenv==1.0.0